src/requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
confusable-homoglyphs==3.2.0
decorator==4.4.2
Django==3.1.4
django-extensions==3.1.0
django-registration==3.1.1
djangorestframework==3.12.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
Pygments==2.7.3
python-decouple==3.4
pytz==2020.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5

